Error handling for TimeSlots provided.

+++ src/rsvp/resources/controller/TimeSlotController.java
@@ -4,6 +4,7 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -13,8 +14,11 @@ import javafx.util.converter.LocalTimeStringConverter;
 import rsvp.resources.DAO.TimeSlotDAO;
 import rsvp.resources.model.TimeSlot;
 import java.time.LocalTime;
+import java.time.format.DateTimeParseException;
 
 public class TimeSlotController {
+    private static final String NO_TIMESLOT_SELECTED_ALERT = "You have to select some time slot in order to do modification";
+    private static final String IMPROPER_HOUR_FORMAT = "You have to provide valid hour format (hh:mm)";
 
     @FXML
     TableView<TimeSlot> timeSlotListTableView;
@@ -40,8 +44,14 @@ public class TimeSlotController {
     ObservableList<TimeSlot> items;
     TimeSlotDAO timeSlotDAO;
 
+    Alert errorAlert;
+
     @FXML
     private void initialize() {
+        errorAlert = new Alert(Alert.AlertType.ERROR);
+        errorAlert.setTitle("Error!");
+        errorAlert.setHeaderText("Error while modyfying data");
+
         items = FXCollections.observableArrayList();
         timeSlotDAO = new TimeSlotDAO();
 
@@ -57,8 +67,15 @@ public class TimeSlotController {
 
     @FXML
     private void handleCreateButtonAction(ActionEvent event) {
-        LocalTime startTime = LocalTime.parse(startTimeFieldCreate.getText());
-        LocalTime endTime = LocalTime.parse(endTimeFieldCreate.getText());
+        LocalTime startTime, endTime;
+        try {
+            startTime = LocalTime.parse(startTimeFieldCreate.getText());
+            endTime = LocalTime.parse(endTimeFieldCreate.getText());
+        } catch(DateTimeParseException e){
+            errorAlert.setContentText(IMPROPER_HOUR_FORMAT);
+            errorAlert.showAndWait();
+            return;
+        }
 
         TimeSlot createdTimeSlot = new TimeSlot(startTime, endTime);
         items.add(createdTimeSlot);
@@ -70,6 +87,11 @@ public class TimeSlotController {
     @FXML
     private void handleDeleteButtonAction(ActionEvent event) {
         TimeSlot chosenTimeSlot = timeSlotListTableView.getSelectionModel().getSelectedItem();
+        if (chosenTimeSlot == null){
+            errorAlert.setContentText(NO_TIMESLOT_SELECTED_ALERT);
+            errorAlert.showAndWait();
+            return;
+        }
         timeSlotDAO.delete(chosenTimeSlot);
         items.remove(chosenTimeSlot);
     }
@@ -79,6 +101,12 @@ public class TimeSlotController {
         LocalTime newStartTime = LocalTime.parse(startTimeFieldUpdate.getText());
         LocalTime newEndTime = LocalTime.parse(endTimeFieldUpdate.getText());
         TimeSlot chosenTimeSlot = timeSlotListTableView.getSelectionModel().getSelectedItem();
+
+        if (chosenTimeSlot == null){
+            errorAlert.setContentText(NO_TIMESLOT_SELECTED_ALERT);
+            errorAlert.showAndWait();
+            return;
+        }
 
         if(!chosenTimeSlot.getStartTime().equals(newStartTime) || !chosenTimeSlot.getEndTime().equals(newEndTime)) {
             timeSlotDAO.update(chosenTimeSlot, newStartTime, newEndTime);
